requirements.txt
Flask==2.1.2
Flask-WTF==0.15.1
mysql-connector-python==8.0.27
requests==2.26.0